expense_tracking_project_2/requirements.txt
streamlit==1.30.0
pandas==2.1.4
plotly==5.9.0
pytest==7.4.0
pydantic==1.10.12
uvicorn==0.35.0
requests==2.31.0
fastapi==0.116.1
mysql-connector-python==9.4.0